Fixed issue with processing

+++ processing/create_database.py
@@ -7,9 +7,9 @@ c.execute('''
     CREATE TABLE stats (
         id INTEGER PRIMARY KEY ASC,
         num_tp_readings INTEGER NOT NULL,
-        max_tp_readings INTEGER,
+        max_tp_readings REAL,
         num_tu_readings INTEGER NOT NULL,
-        max_tu_readings INTEGER,
+        max_tu_readings REAL,
         last_updated VARCHAR(100) NOT NULL
     )
 ''')
