+++ processing/stats.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime
+from sqlalchemy import Column, Integer, String, DateTime, Float
 from base import Base
 
 class Stats(Base):
@@ -8,8 +8,8 @@ class Stats(Base):
     id = Column(Integer, primary_key=True)
     num_tp_readings = Column(Integer, nullable=False)
     num_tu_readings = Column(Integer, nullable=False)
-    max_tp_readings = Column(Integer, nullable=True)
-    max_tu_readings = Column(Integer, nullable=True)
+    max_tp_readings = Column(Float, nullable=True)
+    max_tu_readings = Column(Float, nullable=True)
     last_updated = Column(DateTime, nullable=False)
 
     def __init__(self, num_tp_readings, num_tu_readings,
